Made navbar white and updated nav buttons
I made the background of the navbar white so the font over dark background issue is no longer a problem. I also got rid of the privacy nav button from the template and added ones for all of the sections of the website so the user can easily navigate between them.